add triangle, except... not.
The sine is wavetable generation, with calculations for frequency and
amplitude that make sense.

The triangle attempts to use the onboard triangle generation.  And it
works, but a) the amplitude is ~not even close (you only have 4 bits of
amplitude to use) and b) the frequency is wildly irrelevant.  Because of
the way the generation works, the frequency is actually dependent not
just on the trigger rate, but also the amplitude.  So... demo, but not
for any real use.